+++ CHANGELOG.md
@@ -1,6 +1,22 @@
+# Cozy Improvements 1.2.0 
+**Improvements**  
+- Added configs!
+- Added accessibility improvements
+ - Launch Lever trigger area is much larger (On by default)
+ - Teleporter Buttons can be made larger (Off by default)
+ - Hangar Door buttons can be made larger (Off by default)
+ - Monitor buttons (Switch/Toggle) can be made larger (Off by default)
+
+**Bugfixes**  
+- Redesigned how the Terminal is setup, should permanently fix the issues with it.
+
+
+
 # Cozy Improvements 1.1.0 
 **Bugfixes**  
 - Fixed features to run on all clients instead of just the Host; ACTUALLY Client side now
+
+
 
 # Cozy Improvements 1.0.1  
 **Bugfixes**